IR: Separate out `Intrinsic` node types into separate source
